requirements.txt
PyOpenGL
glfw
numpy
imageio
imgui[glfw]
